move Catalogos components into directory
